requirements.txt
paddlepaddle==3.0.0
paddleocr==3.0.1
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0